Update STM core.

+++ cores/arduino/wiring_digital.cpp
@@ -85,22 +85,22 @@ void pinModeDuet(uint32_t ulPin, enum PinMode ulMode, uint32_t debounceCutoff) {
 }
 
 void digitalWrite(uint32_t ulPin, uint32_t ulVal) {
-  if (ulPin == NC) return;
+  if (ulPin == (uint32_t)NC) return;
   digitalWriteFast(digitalPinToPinName(ulPin), ulVal);
 }
 
 int digitalRead(uint32_t ulPin) {
-  if (ulPin == NC) return 0;
+  if (ulPin == (uint32_t)NC) return 0;
   return digitalReadFast(digitalPinToPinName(ulPin));
 }
 
 void digitalToggle(uint32_t ulPin) {
-  if (ulPin == NC) return;
+  if (ulPin == (uint32_t)NC) return;
   digitalToggleFast(digitalPinToPinName(ulPin));
 }
 
 void setPullup(uint32_t pin, bool en) {
-  if (pin == NC) return;
+  if (pin == (uint32_t)NC) return;
   if (en)
     pin_function(digitalPinToPinName(pin), STM_PIN_DATA(STM_MODE_INPUT, GPIO_PULLUP, 0));
   else
